+++ tests/test_gridresultadmin.py
@@ -10,6 +10,8 @@ import numpy as np
 
 
 from threedigrid.admin.gridresultadmin import GridH5ResultAdmin
+from threedigrid.admin.nodes.models import Nodes
+
 
 test_file_dir = os.path.join(
     os.getcwd(), "tests/test_files")
@@ -144,6 +146,17 @@ def test_timestamps(gr):
 def test_dt_timestamps(gr):
     n_qs = gr.nodes.timeseries(start_time=0, end_time=500)
     assert len(n_qs.dt_timestamps) == len(n_qs.timestamps)
+
+
+def test_get_model_instance_by_field_name(gr):
+    inst = gr.get_model_instance_by_field_name('s1')
+    assert isinstance(inst, Nodes)
+
+
+def test_get_model_instance_by_field_name_raises_index_error(gr):
+    with pytest.raises(IndexError):
+        gr.get_model_instance_by_field_name('zoom_category')
+
 
 # commented for now until the new aggregate.nc is finished